Improve CLI and add check_dae (#310)
This PR improves the CLI experience by having only one command `tqec`
that accept multiple subcommands. It also implements the `tqec
check_dae` subcommand.

Examples:
```
>>> tqec --help
usage: tqec [-h] {dae2observables,check_dae} ...

The main tqec command-line tool.

options:
  -h, --help            show this help message and exit

subcommands:
  {dae2observables,check_dae}
```

```
>>> tqec check_dae --help
usage: tqec check_dae [-h] dae_file

Takes a .dae file in and check that it is correctly recognized by the tqec library.

positional arguments:
  dae_file    The .dae file to check.

options:
  -h, --help  show this help message and exit
```

```
>>> tqec check_dae notebooks/assets/clean_exportable_cnot.dae 
No issue found, the provided .dae file seems valid.
```

The `tqec dae2observables` subcommand should be equivalent to the
previous `dae2observable` command line.
